Add files via upload
Refactored to accommodate variable dimension numbers, so instead of (x, y) there's Point, which can vary from 1 dimension: (x) to as many as your heart desires. Not finished yet.

Also have an old version of the main file and an old newer version which uses Love2d.

Might do further refactoring to see if it's easier to take a more Class-less approach. Remember, this is also a learning process, so a lot of my commits and workflow will look disorganized and amateur.